add checkFieldErrors to EditableTable
